backend/requirements.txt
fastapi
fastapi[test]
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
pydantic>=2.0
pydantic[email]
python-multipart
openpyxl
pytest
httpx
nepali-datetime
cryptography
openai>=1.0.0
google-genai>=1.51.0
